Upgraded ComponentSelect to Angular. #2846
