+++ src/main/webapp/site/src/app/classroom-monitor/component-select/component-select.component.scss
@@ -1,0 +1,7 @@
+mat-select {
+  min-width: 150px;
+}
+
+mat-select, mat-select-trigger {
+  text-transform: none;
+}
