+++ src/main/webapp/site/src/messages.xlf
@@ -5968,6 +5968,24 @@
           <context context-type="linenumber">36</context>
         </context-group>
       </trans-unit>
+      <trans-unit datatype="html" id="92159486420caf44d25e4f85cb0ed4f105d9c854">
+        <source>Assessment items to show</source>
+        <context-group purpose="location">
+          <context context-type="sourcefile">app/classroom-monitor/component-select/component-select.component.html</context>
+          <context context-type="linenumber">8</context>
+        </context-group>
+        <context-group purpose="location">
+          <context context-type="sourcefile">app/classroom-monitor/component-select/component-select.component.html</context>
+          <context context-type="linenumber">16</context>
+        </context-group>
+      </trans-unit>
+      <trans-unit datatype="html" id="6b8ecc745edb25da9e81cee606a8cae77549dde2">
+        <source> 1 assessment item </source>
+        <context-group purpose="location">
+          <context context-type="sourcefile">app/classroom-monitor/component-select/component-select.component.html</context>
+          <context context-type="linenumber">19</context>
+        </context-group>
+      </trans-unit>
       <trans-unit datatype="html" id="7affaae0e14529c0192d5a4bfabd6fcb1ea27529">
         <source>Rubric</source>
         <context-group purpose="location">
