+++ src/main/webapp/wise5/classroomMonitor/classroomMonitorComponents/nodeGrading/nodeGrading.ts
@@ -1,7 +1,7 @@
 'use strict';
 
+import { ComponentSelectComponent } from '../../../../site/src/app/classroom-monitor/component-select/component-select.component';
 import NodeGradingView from './nodeGradingView/nodeGradingView';
-import ComponentSelect from './componentSelect/componentSelect';
 import StepTools from './stepTools/stepTools';
 import WorkgroupItem from './workgroupItem/workgroupItem';
 import * as angular from 'angular';
@@ -11,7 +11,8 @@ import { downgradeComponent } from '@angular/upgrade/static';
 const NodeGrading = angular
   .module('nodeGrading', [])
   .component('nodeGradingView', NodeGradingView)
-  .component('componentSelect', ComponentSelect)
+  .directive('componentSelect',
+      downgradeComponent({ component: ComponentSelectComponent }) as angular.IDirectiveFactory)
   .component('cmStepTools', StepTools)
   .directive('workgroupInfo',
       downgradeComponent({ component: WorkgroupInfoComponent }) as angular.IDirectiveFactory)
