+++ src/main/webapp/wise5/classroomMonitor/classroomMonitorComponents/nodeGrading/nodeGradingView/nodeGradingView.html
@@ -13,8 +13,9 @@
     </div>
     <span flex></span>
     <div class="content-head__more" layout="row" layout-align="center center" layout-wrap>
-      <component-select node-id="{{ ::$ctrl.nodeId }}" custom-class="'md-no-underline md-button'"
-                        on-change="$ctrl.onUpdateHiddenComponents(value)"></component-select>
+      <component-select node-id="{{::$ctrl.nodeId}}"
+                        class="md-button md-no-underline"
+                        (model-change)="$ctrl.onUpdateHiddenComponents($event)"></component-select>
       <md-button ng-if="$ctrl.milestoneReport" ng-click="$ctrl.showReport($event)"
                  class="md-raised info">
         <md-icon class="info"> flag </md-icon>
